refactor ReconnectionManager as interface, allow ClientBuilder to specify reconnectionManager.
